♻️ 서버 컴포넌트 분리로 테스트 메인페이지 데이터를 서버에서 fetching 한다
